refactor: code to use CardEmpty component
This commit refactors the code in two files to replace the FavoriteCardEmpty and ProjectCardEmpty components with a more generic CardEmpty component. The new component is used to display messages when no projects or favorite projects are found, as well as when no snaps are found for a project. The view_size property is also added to the CardEmpty component for better styling.