Update dsign-components-v2 to version 0.15.22 and add new actors to auth_client in upsert page
This commit updates the package.json file by changing the version of dsign-components-v2 from 0.15.17 to 0.15.22, and adds new actors (actor_assets_file_staging, actor_assets_img_staging, and auth_snap_main) to auth_client in the upsert page's +page.svelte file. Additionally, it modifies the handleAddImages function to accept an array of image objects with dataUrl and mimeType properties instead of an array of base64-encoded images, and creates a new async function called commitImgAssetsToStaging that commits image assets to staging using these new actors. Finally, it modifies the handlePublish function to use this new function when publishing a snap with a given name and set of images.