add: cover image selection
This commit updates the `dsign-components-v2` package to version `0.15.24`. It also replaces the `lodash` function `set()` with `findIndex()`, adds a new variable called `cover_img`, and creates a new function called `handleSelectCover()` that sets the value of this variable when an image is selected as the cover for a snap. Finally, it adds a new boolean variable called `is_publishing` which is used to disable certain UI elements while publishing a snap.