add: file_asset object to create_snap_args if file_chunks is not empty
This commit adds a new file_asset object to the create_snap_args object only if the file_chunks variable is not empty. The new file_asset object contains properties such as is_public, content_type and chunk_ids which are used in creating a snap. This change ensures that snaps can be created with files attached only when there are files available.