feat: Update package dependencies and fix snap creation
The `dsign-components-v2` dependency was updated to version `0.17.3`. The SnapCreationModal.svelte file was removed from the project. A console.log statement was removed from the snap page component. The create_snap_args object in the upsert page now has an array for file_asset instead of an empty object when there are no file chunks.