refactor: modals to use dsign-components-v2
This commit refactors the modals in the codebase to use components from the dsign-components-v2 package. The AccountSettingsModal now imports AccountSettings and Modal, while ProjectCreationModal has been deleted and replaced with ProjectUpsertModal. The ProjectDeleteModal now uses ProjectDelete and Modal, while ProjectRenameModal has been deleted. Additionally, modal_mode is used to determine whether a project should be created or edited when opening the ProjectUpsertModal.