refactor: code to remove console logs and add project reference in Snap class
This commit refactors the code by removing unnecessary console logs from two functions in +page.svelte file. It also adds a new field `project_ref` in the Snap class, which is used to store the project id and canister id of a snap. This change is made in three files: Snap.mo, SnapMain.mo, and types.mo. In Snap.mo, the `create_snap` function now takes an argument named `snap_info` instead of `args`. The values for title, image_cover_location, and project_ref are now taken from this argument instead of args.