feat: Update dependencies and add new functionality
This commit updates the dependencies in `dfx.json` and adds a new dependency called "explore". In addition, it introduces a new version of the `Explore` actor with additional functionality to update projects. The `Snap` actor has also been updated to include a reference to the `Explore` actor and call its new function when editing snaps. Finally, there are some minor changes in test files.