refactor: remove `get_profile` and change `get_profile_by_username`
We don't need the call without username, since the one with username will serve both purposes. We also now return a diff type that has the array of projects and favorites. Before it would return `Profile` type which just includes the ids. Also new field `is_owner` will allow the UI to know if it should display buttons to edit vs not.