refactor: project addition to profile
The code has been refactored to streamline the process of adding a project to a user's profile. Previously, an intermediate variable was used to store the updated list of projects before assigning it back to the profile. This step has been eliminated, and now the project is directly added into the user's profile.