refactor: import statements and create new data store
In this commit, we've made several changes to improve the organization of our code. We've refactored the import statements in AccountSettingsModal.svelte and ProjectDeleteModal.svelte for better readability. In +page.svelte, we've commented out calls to projects_update.update_project() and replaced fetch_store.js with data_explore.js in the import statement. Lastly, a new file named data_explore.js has been created which exports a writable Svelte store initialized with an empty project object.