add: 'explore' as a new dependency and refactored deployment script
In this commit, we've added 'explore' as a new dependency in the dfx.json file. We also made significant changes to the deploy_local.sh script, including renaming USERNAME_REGISTRY_PRINCIPAL to USERNAME_REGISTRY_CANISTER_ID for clarity and reordering some of the deployment steps.

We've also made modifications to several Motoko actors. In Creator.mo, we replaced instances of 'this' with 'self'. In Explore.mo, we changed Explore from an actor class to an actor and introduced optional Principal type for username_registry.

Lastly, in UsernameRegistry.mo, we removed set_explore_canister_id function and updated how canister info is saved in Explore Actor.