fix: deployment logs and removed redundant message
In this commit, we've made some changes to improve the clarity of our deployment logs. We've removed a redundant log message that was previously displayed during the installation of child canisters. Additionally, we've enhanced the information provided in the deployment log messages by including both the name and ID of each deployed canister for easier identification and tracking.