add: tag update functionality
A new function `update_tags` has been introduced to handle the updating of tags. This function is triggered by an event and uses the details from this event to update the snap tags. It also handles error scenarios, logging any issues encountered during the process. The function is now linked with the `update_tags` event in the UI.