refactor: project deletion and update logging
The code changes refactor the process of deleting a project. The filtering out of the project ID to be deleted has been moved to a utility function for better code organization. Additionally, the deletion process now includes removing associated snaps.

Logging in the UsernameRegistry actor has also been updated. Previously commented-out logging lines have been reinstated, and new logs have been added for when maximum users are exceeded or a user is added to a new creator canister.

Lastly, the version number in UsernameRegistry has been incremented from 7 to 8.